requirements.txt
Django==5.1
django-rest-framework==0.1.0
drf-spectacular==0.28.0
django-extensions==4.1